+++ controllers/ControllerDecorator.ts
@@ -15,6 +15,7 @@ import {
 import {Param, PARAMS_KEY, ParamType} from '../params/ParamDecorators';
 import {ErrorHandlerManager, ERRORHANDLER_KEY, DEFAULT_ERROR_HANDLER} from '../errors/ErrorHandlerDecorator';
 import {Validator} from '../validators/Validators';
+import {RequestHandler} from 'express-serve-static-core';
 import httpStatus = require('http-status');
 
 let controllers: ControllerRegistration[] = [],
@@ -32,7 +33,7 @@ try {
 }
 
 class ControllerRegistration {
-    constructor(public controller: any, public prefix?: string) {
+    constructor(public controller: any, public prefix?: string, public middlewares: RequestHandler[] = []) {
     }
 }
 
@@ -103,22 +104,22 @@ function getParamValues(params: Param[], request: Request, response: Response) {
     return paramValues;
 }
 
-function registerRoute(route: RouteRegistration, router: Router, routeUrl: string) {
+function registerRoute(route: RouteRegistration, router: Router, routeUrl: string, middlewares: RequestHandler[]) {
     switch (route.method) {
         case RouteMethod.Get:
-            router.get(routeUrl, (route.descriptor.value as any));
+            router.get(routeUrl, ...middlewares, (route.descriptor.value as any));
             break;
         case RouteMethod.Put:
-            router.put(routeUrl, (route.descriptor.value as any));
+            router.put(routeUrl, ...middlewares, (route.descriptor.value as any));
             break;
         case RouteMethod.Post:
-            router.post(routeUrl, (route.descriptor.value as any));
+            router.post(routeUrl, ...middlewares, (route.descriptor.value as any));
             break;
         case RouteMethod.Delete:
-            router.delete(routeUrl, (route.descriptor.value as any));
+            router.delete(routeUrl, ...middlewares, (route.descriptor.value as any));
             break;
         case RouteMethod.Head:
-            router.head(routeUrl, (route.descriptor.value as any));
+            router.head(routeUrl, ...middlewares, (route.descriptor.value as any));
             break;
         default:
             throw new HttpVerbNotSupportedError(route.method);
@@ -130,11 +131,12 @@ function registerRoute(route: RouteRegistration, router: Router, routeUrl: strin
  * expressJS router when "registerControllers" is called.
  *
  * @param {string} [routePrefix] - Prefix for the whole controller. This path is added to all routes.
+ * @param {RequestHandler[]} [middlewares] - Middleware functions for the controller to be executed before the routing functions.
  * @returns {(Function) => void} - Decorator for the controller class.
  */
-export function Controller(routePrefix?: string) {
+export function Controller(routePrefix?: string, ...middlewares: RequestHandler[]) {
     return (controller: any) => {
-        controllers.push(new ControllerRegistration(controller, routePrefix));
+        controllers.push(new ControllerRegistration(controller, routePrefix, middlewares));
     };
 }
 
@@ -166,6 +168,7 @@ export function registerControllers(baseUrl: string = '', router: Router = Route
                 routeId = routeUrl + route.method.toString(),
                 returnType = Reflect.getMetadata('design:returntype', routeTarget, route.propertyKey),
                 params: Param[] = Reflect.getOwnMetadata(PARAMS_KEY, routeTarget, route.propertyKey) || [],
+                middlewares = [...ctrl.middlewares, ...route.middlewares],
                 method = route.descriptor.value,
                 hasResponseParam = !!params.filter(p => p.paramType === ParamType.Response).length;
 
@@ -242,7 +245,7 @@ export function registerControllers(baseUrl: string = '', router: Router = Route
                 }
             };
 
-            registerRoute(route, router, routeUrl);
+            registerRoute(route, router, routeUrl, middlewares);
             definedRoutes.push(routeId);
         });
     });